+++ skore/src/skore/_utils/_callable.py
